+++ web/src/actions/apd.test.js
@@ -4,7 +4,7 @@ import thunk from 'redux-thunk';
 import sinon from 'sinon';
 
 import * as actions from './apd';
-import * as appActions from './app';
+import { SELECT_APD } from './app';
 import { ARIA_ANNOUNCE_CHANGE } from './aria';
 import { UPDATE_BUDGET } from './budget';
 import axios from '../util/api';
@@ -36,18 +36,6 @@ describe('apd actions', () => {
     });
   });
 
-  it('requestSave should create SAVE_APD_REQUEST action', () => {
-    expect(actions.requestSave()).toEqual({ type: actions.SAVE_APD_REQUEST });
-  });
-
-  it('saveSuccess should create SAVE_APD_SUCCESS action', () => {
-    expect(actions.saveSuccess()).toEqual({ type: actions.SAVE_APD_SUCCESS });
-  });
-
-  it('saveFailure should create SAVE_APD_FAILURE action', () => {
-    expect(actions.saveFailure()).toEqual({ type: actions.SAVE_APD_FAILURE });
-  });
-
   it('selectApdOnLoad should create SET_SELECT_APD_ON_LOAD action if user is not an admin', async () => {
     const store = mockStore({ user: { data: { role: 'not an admin' } } });
 
@@ -66,47 +54,6 @@ describe('apd actions', () => {
     await store.dispatch(actions.selectApdOnLoad());
 
     expect(store.getActions()).toEqual([]);
-  });
-
-  it('selectAPD should create SELECT_APD action, redirect to provided route, and save APD ID to local storage', async () => {
-    const apd = {
-      id: 'apd-id',
-      selected: 'apd goes here',
-      federalCitations: { already: 'exists' }
-    };
-    fetchMock.onGet('/apds/apd-id').reply(200, apd);
-
-    const deserialize = sinon.stub().returns('deserialized apd');
-
-    const state = { apd: { byId: { apdID: 'hello there' } } };
-    const store = mockStore(state);
-    const testRoute = '/test';
-
-    const global = { localStorage: { setItem: sinon.stub() } };
-    const pushRoute = route => ({ type: 'FAKE_PUSH', pushRoute: route });
-
-    const expectedActions = [
-      { type: ARIA_ANNOUNCE_CHANGE, message: 'Your APD is loading' },
-      { type: appActions.SELECT_APD, apd: 'deserialized apd' },
-      { type: UPDATE_BUDGET, state },
-      { type: 'FAKE_PUSH', pushRoute: testRoute },
-      {
-        type: ARIA_ANNOUNCE_CHANGE,
-        message: 'Your APD is loaded and ready to edit'
-      }
-    ];
-
-    await store.dispatch(
-      appActions.selectApd('apd-id', '/test', {
-        deserialize,
-        global,
-        pushRoute
-      })
-    );
-
-    expect(store.getActions()).toEqual(expectedActions);
-    expect(global.localStorage.setItem.calledWith('last-apd-id', 'apdID'));
-    expect(deserialize.calledWith(apd)).toEqual(true);
   });
 
   it('createRequest should create CREATE_APD_REQUEST action', () => {
@@ -152,7 +99,7 @@ describe('apd actions', () => {
         { type: actions.CREATE_APD_REQUEST },
         { type: actions.CREATE_APD_SUCCESS, data: apd },
         { type: ARIA_ANNOUNCE_CHANGE, message: 'Your APD is loading' },
-        { type: appActions.SELECT_APD, apd },
+        { type: SELECT_APD, apd },
         { type: UPDATE_BUDGET, state },
         { type: 'FAKE_PUSH', pushRoute: '/apd' },
         {
@@ -345,141 +292,6 @@ describe('apd actions', () => {
 
       await store.dispatch(actions.fetchApdDataIfNeeded());
       expect(store.getActions()).toEqual([]);
-    });
-  });
-
-  describe('save APD to API', () => {
-    const serialize = sinon.mock();
-    const serializedApd = {
-      activities: [
-        {
-          alternatives: 'alternatives approach',
-          contractorResources: 'contractors',
-          costAllocation: 'give us those dollars',
-          costAllocationNarrative: 'cost allocation narrative',
-          description: 'activity description',
-          expenses: 'paper, pens, airplanes, and bouncy castles',
-          fundingSource: 'funding source',
-          goals: 'build the best Medicaid IT system ever seen',
-          id: 'activity 1',
-          key: 'activity 1',
-          name: 'activity name',
-          schedule: 'before the heat death of the universe',
-          standardsAndConditions: 'florp',
-          statePersonnel: 'the people who work here',
-          summary: 'activity summary',
-          quarterlyFFP: 'we want money a little at a time'
-        },
-        {
-          alternatives: 'alternatives approach 2',
-          costAllocationDesc: 'cost allocation methodology 2',
-          description: 'activity description 2',
-          fundingSource: 'funding source 2',
-          id: 'activity 2',
-          key: 'activity 2',
-          name: 'activity 2 name',
-          otherFundingDesc: 'other funding sources 2',
-          summary: 'activity summary 2'
-        }
-      ],
-      federalCitations: 'CFR 395.2362.472462.2352.36 (b) three',
-      id: 'id-to-update',
-      incentivePayments: 'money to do good work',
-      narrativeHIE: 'HIE narrative text',
-      narrativeHIT: 'HIT narrative text',
-      narrativeMMIS: 'MMIS narrative text',
-      programOverview: 'APD overview text',
-      pointsOfContact: 'people to call if stuff goes sour',
-      previousActivityExpenses: 'money we spent last time',
-      previousActivitySummary: 'other activities happened in the past',
-      stateProfile: 'we like long walks on the beach',
-      summary: 'apd summary',
-      years: ['1992', '1993']
-    };
-
-    const state = {
-      notification: { open: false, queue: [] },
-      apd: {
-        data: {
-          id: 'id-to-update'
-        }
-      },
-      activities: {},
-      patch: []
-    };
-
-    beforeEach(() => {
-      serialize.resetBehavior();
-      serialize.resetHistory();
-
-      // Create a new cloned object for each test because
-      // the APD action mutates the returned value.  If
-      // we don't reset, each test begins with the results
-      // of the previous test's mutations. 😬
-      serialize.returns(JSON.parse(JSON.stringify(serializedApd)));
-
-      state.patch = ['these', 'get', 'sent', 'off'];
-
-      fetchMock.reset();
-    });
-
-    it('creates save request but does not actually send save if not dirty', () => {
-      state.patch.length = 0;
-      const store = mockStore(state);
-      fetchMock.onPatch('/apds/id-to-update').reply(403, [{ foo: 'bar' }]);
-
-      const expectedActions = [{ type: actions.SAVE_APD_REQUEST }];
-
-      return store.dispatch(actions.saveApd({ serialize })).catch(() => {
-        expect(store.getActions()).toEqual(expectedActions);
-        expect(serialize.calledWith(state.apd.data, state.activities)).toEqual(
-          true
-        );
-      });
-    });
-
-    it('creates save request and logged-out actions if the user is not logged in', () => {
-      const store = mockStore(state);
-      fetchMock.onPatch('/apds/id-to-update').reply(403, [{ foo: 'bar' }]);
-
-      const expectedActions = [
-        { type: actions.SAVE_APD_REQUEST },
-        { type: actions.SAVE_APD_FAILURE, data: 'save-apd.not-logged-in' }
-      ];
-
-      return store.dispatch(actions.saveApd()).catch(() => {
-        expect(store.getActions()).toEqual(expectedActions);
-      });
-    });
-
-    it('creates save request and save failure actions if the save fails', () => {
-      const store = mockStore(state);
-      fetchMock.onPatch('/apds/id-to-update').reply(400, [{ foo: 'bar' }]);
-
-      const expectedActions = [
-        { type: actions.SAVE_APD_REQUEST },
-        { type: actions.SAVE_APD_FAILURE }
-      ];
-
-      return store.dispatch(actions.saveApd()).catch(() => {
-        expect(store.getActions()).toEqual(expectedActions);
-      });
-    });
-
-    it('saves and does all the good things', () => {
-      const updatedApd = {};
-      const store = mockStore(state);
-
-      fetchMock.onPatch('/apds/id-to-update').reply(200, updatedApd);
-
-      const expectedActions = [
-        { type: actions.SAVE_APD_REQUEST },
-        { type: actions.SAVE_APD_SUCCESS, data: updatedApd }
-      ];
-
-      return store.dispatch(actions.saveApd()).then(() => {
-        expect(store.getActions()).toEqual(expectedActions);
-      });
     });
   });
 });
