+++ web/src/actions/app/symbols.js
@@ -1,3 +1,5 @@
-export const SELECT_APD = Symbol('select apd');
+export const SAVE_APD_FAILURE = Symbol('save apd failure');
+export const SAVE_APD_REQUEST = Symbol('save apd request');
+export const SAVE_APD_SUCCESS = Symbol('save apd success');
 
-export default { SELECT_APD };
+export const SELECT_APD = Symbol('select apd');
