+++ web/src/reducers/errors.test.js
@@ -21,7 +21,7 @@ import {
   SAVE_APD_FAILURE,
   SAVE_APD_REQUEST,
   SAVE_APD_SUCCESS
-} from '../actions/apd';
+} from '../actions/app';
 
 describe('errors reducer', () => {
   const initialState = {
