+++ web/src/reducers/patch.test.js
@@ -1,4 +1,4 @@
-import { SAVE_APD_SUCCESS } from '../actions/apd';
+import { SAVE_APD_SUCCESS } from '../actions/app';
 import {
   ADD_APD_ITEM,
   ADD_APD_YEAR,
