+++ web/src/reducers/working.test.js
@@ -20,7 +20,7 @@ import {
   SAVE_APD_FAILURE,
   SAVE_APD_REQUEST,
   SAVE_APD_SUCCESS
-} from '../actions/apd';
+} from '../actions/app';
 
 describe('working state reducer', () => {
   const initialState = {
